added separate route for register view
